extract create wiki dialog to component
